Update DFRobot_Oxygen.py
The register 3 in the memory, sometimes produces wrong values over 200 where they should be generally around 85 in the outdoor atmosphere. This causes the output values to be way off and jumpy. I added some lines to skip those values. This also causes a divisionbyzero error in the first run that had to be handled.